+++ src/utility/angle.cpp
@@ -8,11 +8,11 @@
 #include <numbers>
 
 namespace ja {
-    double angle::degrees() const {
+    float angle::degrees() const {
         return m_radians * (180.0 / std::numbers::pi);
     }
 
-    double angle::radians() const {
+    float angle::radians() const {
         return m_radians;
     }
